mass change of Id values from int to long

+++ src/main/java/com/podio/app/AppAPI.java
@@ -29,7 +29,7 @@ public class AppAPI extends BaseAPI {
 	 *            The id of the app to be returned
 	 * @return The full definition of the app
 	 */
-	public Application getApp(int appId) {
+	public Application getApp(long appId) {
 		return getResourceFactory().getApiResource("/app/" + appId).get(
 				Application.class);
 	}
@@ -42,7 +42,7 @@ public class AppAPI extends BaseAPI {
 	 *            The id of the space
 	 * @return The list of apps on the given space
 	 */
-	public List<Application> getAppsOnSpace(int spaceId) {
+	public List<Application> getAppsOnSpace(long spaceId) {
 		return getResourceFactory().getApiResource(
 				"/app/space/" + spaceId + "/").get(
 				new GenericType<List<Application>>() {
@@ -73,7 +73,7 @@ public class AppAPI extends BaseAPI {
 	 *            The definition for the new app
 	 * @return The id of the newly created app
 	 */
-	public int addApp(ApplicationCreate app) {
+	public long addApp(ApplicationCreate app) {
 		return getResourceFactory().getApiResource("/app/")
 				.entity(app, MediaType.APPLICATION_JSON_TYPE)
 				.post(ApplicationCreateResponse.class).getId();
@@ -94,7 +94,7 @@ public class AppAPI extends BaseAPI {
 	 * @param app
 	 *            The updated app definition
 	 */
-	public void updateApp(int appId, ApplicationUpdate app) {
+	public void updateApp(long appId, ApplicationUpdate app) {
 		getResourceFactory().getApiResource("/app/" + appId)
 				.entity(app, MediaType.APPLICATION_JSON).put();
 	}
@@ -108,7 +108,7 @@ public class AppAPI extends BaseAPI {
 	 *            The definition of the new field
 	 * @return The id of the newly created field
 	 */
-	public int addField(int appId, ApplicationFieldCreate field) {
+	public long addField(long appId, ApplicationFieldCreate field) {
 		return getResourceFactory().getApiResource("/app/" + appId + "/field/")
 				.entity(field, MediaType.APPLICATION_JSON_TYPE)
 				.post(ApplicationFieldCreateResponse.class).getId();
@@ -125,7 +125,7 @@ public class AppAPI extends BaseAPI {
 	 * @param configuration
 	 *            The new configuration of the field
 	 */
-	public void updateField(int appId, int fieldId,
+	public void updateField(long appId, long fieldId,
 			ApplicationFieldConfiguration configuration) {
 		getResourceFactory()
 				.getApiResource("/app/" + appId + "/field/" + fieldId)
@@ -141,7 +141,7 @@ public class AppAPI extends BaseAPI {
 	 *            The id of the field to be returned
 	 * @return The definition and current configuration of the requested field
 	 */
-	public ApplicationField getField(int appId, int fieldId) {
+	public ApplicationField getField(long appId, long fieldId) {
 		return getResourceFactory().getApiResource(
 				"/app/" + appId + "/field/" + fieldId).get(
 				ApplicationField.class);
@@ -156,7 +156,7 @@ public class AppAPI extends BaseAPI {
 	 *            The id of the field to be returned
 	 * @return The definition and current configuration of the requested field
 	 */
-	public ApplicationField getField(int appId, String externalId) {
+	public ApplicationField getField(long appId, String externalId) {
 		return getResourceFactory().getApiResource(
 				"/app/" + appId + "/field/" + externalId).get(
 				ApplicationField.class);
@@ -172,7 +172,7 @@ public class AppAPI extends BaseAPI {
 	 * @param fieldId
 	 *            The id of the field that should be deleted
 	 */
-	public void deleteField(int appId, int fieldId) {
+	public void deleteField(long appId, long fieldId) {
 		getResourceFactory().getApiResource(
 				"/app/" + appId + "/field/" + fieldId).delete();
 	}
@@ -186,7 +186,7 @@ public class AppAPI extends BaseAPI {
 	 *            The id of the space the app should be installed o n
 	 * @return The id of the newly installed app
 	 */
-	public int install(int appId, int spaceId) {
+	public long install(long appId, long spaceId) {
 		return getResourceFactory()
 				.getApiResource("/app/" + appId + "/install")
 				.entity(new ApplicationInstall(spaceId),
@@ -203,7 +203,7 @@ public class AppAPI extends BaseAPI {
 	 * @param appIds
 	 *            The ids of the apps in the new order
 	 */
-	public void updateOrder(int spaceId, List<Integer> appIds) {
+	public void updateOrder(long spaceId, List<Integer> appIds) {
 		getResourceFactory().getApiResource("/app/space/" + spaceId + "/order")
 				.entity(appIds, MediaType.APPLICATION_JSON_TYPE).put();
 	}
@@ -226,7 +226,7 @@ public class AppAPI extends BaseAPI {
 	 *            The id of the app the dependecies should be returned for
 	 * @return The applications that the given app depends on
 	 */
-	public Dependencies getDependencies(int appId) {
+	public Dependencies getDependencies(long appId) {
 		return getResourceFactory().getApiResource(
 				"/app/" + appId + "/dependencies/").get(Dependencies.class);
 	}
@@ -238,7 +238,7 @@ public class AppAPI extends BaseAPI {
 	 * @param appId
 	 *            The id of the app to deactivate
 	 */
-	public void deactivateApp(int appId) {
+	public void deactivateApp(long appId) {
 		getResourceFactory().getApiResource("/app/" + appId + "/deactivate")
 				.entity(new Empty(), MediaType.APPLICATION_JSON_TYPE).post();
 	}
@@ -250,7 +250,7 @@ public class AppAPI extends BaseAPI {
 	 * @param appId
 	 *            The id of the app to activate
 	 */
-	public void activateApp(int appId) {
+	public void activateApp(long appId) {
 		getResourceFactory().getApiResource("/app/" + appId + "/activate")
 				.entity(new Empty(), MediaType.APPLICATION_JSON_TYPE).post();
 	}
@@ -262,7 +262,7 @@ public class AppAPI extends BaseAPI {
 	 * @param appId
 	 *            The id of the app to delete
 	 */
-	public void deleteApp(int appId) {
+	public void deleteApp(long appId) {
 		getResourceFactory().getApiResource("/app/" + appId).delete();
 	}
 }
