+++ src/main/java/com/podio/contact/ProfileBase.java
@@ -11,7 +11,7 @@ public class ProfileBase implements Serializable {
 	/**
 	 * The id of the profile
 	 */
-	private int profileId;
+	private long profileId;
 
 	/**
 	 * The id of the user
@@ -19,7 +19,7 @@ public class ProfileBase implements Serializable {
 	private Integer userId;
 
 	@Override
-	public int hashCode() {
+	public long hashCode() {
 		final int prime = 31;
 		int result = 1;
 		result = prime * result + profileId;
@@ -41,12 +41,12 @@ public class ProfileBase implements Serializable {
 	}
 
 	@JsonProperty("profile_id")
-	public int getProfileId() {
+	public long getProfileId() {
 		return profileId;
 	}
 
 	@JsonProperty("profile_id")
-	public void setProfileId(int profileId) {
+	public void setProfileId(long profileId) {
 		this.profileId = profileId;
 	}
