+++ src/main/java/com/podio/item/FieldValuesUpdate.java
@@ -27,19 +27,19 @@ public class FieldValuesUpdate {
 		super();
 	}
 
-	public FieldValuesUpdate(int id, List<Map<String, ?>> values) {
+	public FieldValuesUpdate(long id, List<Map<String, ?>> values) {
 		super();
 		this.id = id;
 		this.values = values;
 	}
 
-	public FieldValuesUpdate(int id, Map<String, ?> value) {
+	public FieldValuesUpdate(long id, Map<String, ?> value) {
 		super();
 		this.id = id;
 		this.values = Collections.<Map<String, ?>> singletonList(value);
 	}
 
-	public FieldValuesUpdate(int id, String subId, Object value) {
+	public FieldValuesUpdate(long id, String subId, Object value) {
 		super();
 		this.id = id;
 		this.values = Collections.<Map<String, ?>> singletonList(Collections
