+++ src/main/java/com/podio/item/ItemAPI.java
@@ -42,7 +42,7 @@ public class ItemAPI extends BaseAPI {
 	 *            True if the create should be silten, false otherwise
 	 * @return The id of the newly created item
 	 */
-	public int addItem(int appId, ItemCreate create, boolean silent) {
+	public long addItem(long appId, ItemCreate create, boolean silent) {
 		return getResourceFactory().getApiResource("/item/app/" + appId + "/")
 				.queryParam("silent", silent ? "1" : "0")
 				.entity(create, MediaType.APPLICATION_JSON_TYPE)
@@ -56,7 +56,7 @@ public class ItemAPI extends BaseAPI {
 	 *            The id of the item
 	 * @return The item with given id
 	 */
-	public Item getItem(int itemId) {
+	public Item getItem(long itemId) {
 		return getResourceFactory().getApiResource("/item/" + itemId).get(
 				Item.class);
 	}
@@ -75,7 +75,7 @@ public class ItemAPI extends BaseAPI {
 	 * @param hook
 	 *            True if hooks should be executed for the change, false otherwise
 	 */
-	public void updateItem(int itemId, ItemUpdate update, boolean silent, boolean hook) {
+	public void updateItem(long itemId, ItemUpdate update, boolean silent, boolean hook) {
 		getResourceFactory().getApiResource("/item/" + itemId)
 				.queryParam("silent", silent ? "1" : "0")
 				.queryParam("hook", hook ? "1" : "0")
@@ -94,7 +94,7 @@ public class ItemAPI extends BaseAPI {
 	 * @param hook
 	 *            True if hooks should be executed for the change, false otherwise
 	 */
-	public void updateItemValues(int itemId, List<FieldValuesUpdate> values,
+	public void updateItemValues(long itemId, List<FieldValuesUpdate> values,
 			boolean silent, boolean hook) {
 		getResourceFactory().getApiResource("/item/" + itemId + "/value/")
 				.queryParam("silent", silent ? "1" : "0")
@@ -116,7 +116,7 @@ public class ItemAPI extends BaseAPI {
 	 * @param hook
 	 *            True if hooks should be executed for the change, false otherwise
 	 */
-	public void updateItemFieldValues(int itemId, int fieldId,
+	public void updateItemFieldValues(long itemId, long fieldId,
 			List<Map<String, Object>> values, boolean silent, boolean hook) {
 		getResourceFactory()
 				.getApiResource("/item/" + itemId + "/value/" + fieldId)
@@ -134,7 +134,7 @@ public class ItemAPI extends BaseAPI {
 	 * @param silent
 	 *            True if the deletion should be silent, false otherwise
 	 */
-	public void deleteItem(int itemId, boolean silent) {
+	public void deleteItem(long itemId, boolean silent) {
 		getResourceFactory().getApiResource("/item/" + itemId)
 				.queryParam("silent", silent ? "1" : "0").delete();
 	}
@@ -148,7 +148,7 @@ public class ItemAPI extends BaseAPI {
 	 *            The id of the field
 	 * @return The values on the field on the item
 	 */
-	public List<Map<String, Object>> getItemFieldValues(int itemId, int fieldId) {
+	public List<Map<String, Object>> getItemFieldValues(long itemId, long fieldId) {
 		return getResourceFactory().getApiResource(
 				"/item/" + itemId + "/value/" + fieldId).get(
 				new GenericType<List<Map<String, Object>>>() {
@@ -163,7 +163,7 @@ public class ItemAPI extends BaseAPI {
 	 *            The id of the item
 	 * @return The values on the item
 	 */
-	public List<FieldValuesView> getItemValues(int itemId) {
+	public List<FieldValuesView> getItemValues(long itemId) {
 		return getResourceFactory().getApiResource(
 				"/item/" + itemId + "/value/").get(
 				new GenericType<List<FieldValuesView>>() {
@@ -184,7 +184,7 @@ public class ItemAPI extends BaseAPI {
 	 *            The maximum number of results to return. Default value: 13
 	 * @return The items that were valid for the field and with text matching
 	 */
-	public List<ItemMini> getItemsByFieldAndTitle(int fieldId, String text,
+	public List<ItemMini> getItemsByFieldAndTitle(long fieldId, String text,
 												List<Integer> notItemIds, Integer limit) {
 		WebResource resource = getResourceFactory().getApiResource(
 			"/item/field/" + fieldId + "/find");
@@ -210,7 +210,7 @@ public class ItemAPI extends BaseAPI {
 	 *            The id of the item
 	 * @return The references to the given item
 	 */
-	public List<ItemReference> getItemReference(int itemId) {
+	public List<ItemReference> getItemReference(long itemId) {
 		return getResourceFactory().getApiResource(
 				"/item/" + itemId + "/reference/").get(
 				new GenericType<List<ItemReference>>() {
@@ -227,7 +227,7 @@ public class ItemAPI extends BaseAPI {
 	 *            revision
 	 * @return The revision
 	 */
-	public ItemRevision getItemRevision(int itemId, int revisionId) {
+	public ItemRevision getItemRevision(long itemId, long revisionId) {
 		return getResourceFactory().getApiResource(
 				"/item/" + itemId + "/revision/" + revisionId).get(
 				ItemRevision.class);
@@ -244,8 +244,8 @@ public class ItemAPI extends BaseAPI {
 	 *            The to revision
 	 * @return The difference between the two revision
 	 */
-	public List<ItemFieldDifference> getItemRevisionDifference(int itemId,
-			int revisionFrom, int revisionTo) {
+	public List<ItemFieldDifference> getItemRevisionDifference(long itemId,
+			long revisionFrom, long revisionTo) {
 		return getResourceFactory().getApiResource(
 				"/item/" + itemId + "/revision/" + revisionFrom + "/"
 						+ revisionTo).get(
@@ -260,14 +260,14 @@ public class ItemAPI extends BaseAPI {
 	 *            The id of the item
 	 * @return All the revisions
 	 */
-	public List<ItemRevision> getItemRevisions(int itemId) {
+	public List<ItemRevision> getItemRevisions(long itemId) {
 		return getResourceFactory().getApiResource(
 				"/item/" + itemId + "/revision/").get(
 				new GenericType<List<ItemRevision>>() {
 				});
 	}
 
-	public ItemsResponse filterItems(int appId, ItemFilter filter) {
+	public ItemsResponse filterItems(long appId, ItemFilter filter) {
 		return getResourceFactory().getApiResource("/item/app/" + appId + "/filter/")
 				.entity(filter, MediaType.APPLICATION_JSON_TYPE).post(ItemsResponse.class);
 	}
@@ -291,7 +291,7 @@ public class ItemAPI extends BaseAPI {
 	 *            The filters to apply
 	 * @return The items matching the filters
 	 */
-	public ItemsResponse getItems(int appId, Integer limit, Integer offset,
+	public ItemsResponse getItems(long appId, Integer limit, Integer offset,
 			SortBy sortBy, Boolean sortDesc, FilterByValue<?>... filters) {
 		WebResource resource = getResourceFactory().getApiResource(
 				"/item/app/" + appId + "/v2/");
@@ -324,7 +324,7 @@ public class ItemAPI extends BaseAPI {
 	 *            The external id
 	 * @return The items matching the app and external id
 	 */
-	public ItemsResponse getItemsByExternalId(int appId, String externalId) {
+	public ItemsResponse getItemsByExternalId(long appId, String externalId) {
 		return getItems(appId, null, null, null, null,
 				new FilterByValue<String>(new ExternalIdFilterBy(), externalId));
 	}
@@ -350,7 +350,7 @@ public class ItemAPI extends BaseAPI {
 	 *            The filters to apply
 	 * @return The items matching the filters
 	 */
-	public ItemsResponse getItemsByView(int appId, int viewId, Integer limit, Integer offset,
+	public ItemsResponse getItemsByView(long appId, long viewId, Integer limit, Integer offset,
 			SortBy sortBy, Boolean sortDesc, FilterByValue<?>... filters) {
 		WebResource resource = getResourceFactory().getApiResource(
 				"/item/app/" + appId + "/filter/" + viewId + "/");
